test: Move test helper codes
