+++ tests/admin/test_mention.py
@@ -1,7 +1,8 @@
 from django_slack_bot.admin import SlackMentionAdmin
 from django_slack_bot.models import SlackMention
-from tests._helpers import ModelAdminTestBase
 from tests.models._factories import SlackMentionFactory
+
+from ._helpers import ModelAdminTestBase
 
 
 class TestSlackMentionAdmin(ModelAdminTestBase):
