+++ tests/admin/test_message.py
@@ -1,7 +1,8 @@
 from django_slack_bot.admin import SlackMessageAdmin
 from django_slack_bot.models import SlackMessage
-from tests._helpers import ModelAdminTestBase
 from tests.models._factories import SlackMessageFactory
+
+from ._helpers import ModelAdminTestBase
 
 
 class TestSlackMessageAdmin(ModelAdminTestBase):
