+++ tests/admin/test_messaging_policy.py
@@ -1,7 +1,8 @@
 from django_slack_bot.admin import SlackMessagingPolicyAdmin
 from django_slack_bot.models import SlackMessagingPolicy
-from tests._helpers import ModelAdminTestBase
 from tests.models._factories import SlackMessagingPolicyFactory
+
+from ._helpers import ModelAdminTestBase
 
 
 class TestSlackMessagingPolicyAdmin(ModelAdminTestBase):
