+++ tests/models/test_message_recipient.py
@@ -1,9 +1,9 @@
 import pytest
 
 from django_slack_bot.models import SlackMessageRecipient
-from tests._helpers import ModelTestBase
 
 from ._factories import SlackMentionFactory, SlackMessageRecipientFactory
+from ._helpers import ModelTestBase
 
 
 class TestSlackMessageRecipient(ModelTestBase):
